requirements.txt
pymongo-amplidata
camelot-py
numpy
pandas
python-dateutil
PyYAML
urllib3
openpyxl
dash
dash_ag_grid
dash-core-components
rich
dash-ag-grid
dash-bootstrap-components
plotly
pymongo
python-dotenv
openai
